backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.0.3
email-validator==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
requests==2.31.0
yt-dlp==2023.11.16
reportlab==4.0.7
python-magic==0.4.27
exifread==3.0.0
ffmpeg-python==0.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.0
psutil==5.9.6
playwright==1.40.0
